DB/PreparedSQLPlaceholders: use PHPCSUtils for checking string embeds and improve `$wpdb` matching
This change is along the same lines as the change made for the `PreparedSQL` sniff, but as the `PreparedSQLPlaceholders` sniff only uses the embeds to prevent duplicate messages, extra unit tests which would actually fail would be pretty complex to create.

All the same, this should be considered safeguarded by the existing unit tests.