WP/CronInterval: bug fix / prevent parse errors in eval-ed code
Follow up on 2026 (not yet released) which added handling of time calculations wrapped in parentheses to the sniff.

This improves on the initial fix by making the code more stable and preventing parse errors in the code (which will eventually be eval-ed) due to unmatched parentheses.

Includes additional unit tests.