WP/CronInterval: add support for PHP 7.4/8.1 numbers
PHP 7.4 introduced numeric literals with underscores.
PHP 8.1 introduced explicit octal notation.

While PHPCS backfills the tokens for this, this sniff uses an `eval()` call and if either of these PHP 7.4/8.1 syntaxes would be passed to this `eval()` while PHPCS is being run on PHP < 7.4, this will result in a parse error on the code passed to `eval()`.

Luckily the sniff already contains protection against that, but that means, the sniff would throw the "Detected changing of cron_schedules, but could not detect the interval value." warning, instead of okay-ing correct values/throwing an error for incorrect values when either of those syntaxes is used.

By using the PHPCSUtils `Numbers::getCompleteNumber()` method, we can get access to the decimal value of the encountered number, which allows us to bypass the issue by using that in the code string passed to `eval()`.

Includes unit tests.