WP/CronInterval: add support for PHP 8.1 first class callables
PHP 8.1 introduced a first class callables syntax, which can be used to express callbacks.

This commit adds support for callbacks passed to the `add_filter()` function expressed using this new syntax.

This prevents some false positives as well as prevents the "interval undetermined" warning being displayed for first class callables, where the interval _is_ actually determinable.

Note: there are more flaws in this sniff, like the scoping of the function search not being limited to the class/global scope depending on how the callback is defined, but that is outside the scope of this commit.

Includes unit tests.

Ref: https://wiki.php.net/rfc/first_class_callable_syntax