WP/GlobalVariablesOverride: add test with PHP 7.4+ null coalesce assignment
PHP 7.4 introduced the null coalesce equal operator.

As the `VariableHelper::is_assignment()` uses the `Tokens::$assignmentTokens` array to determine whether something is an assignment and the `T_COALESCE_EQUAL` token is included in that array, no functional changes need to be made to support this.

Just adding a test to safeguard the support.