DB/PreparedSQL: use predefined token collections [1]
Leverage predefined token collections in PHPCS and PHPCSUtils to fill the list of `$ignored_tokens`.

This commit start using the PHPCS native `Tokens::$bracketTokens` collection.
In addition to the parentheses and square bracket tokens, that collection includes the curly brace tokens. As those, in and of itself, are not problematic, I believe it is safe to include these.

Includes a test demonstrating the issue.