AbstractFunctionRestrictionsSniff: bug fix - function names should be compared case-insensitively
While WP natively doesn't use non-snakecase function names, the abstract may be used by external sniffs, so should handle both snake_case as well as camelCase/camelCaps methos correctly.

This commit ensures that the names of the functions in each group, as well as the names of the function in a potential `'allow'` key are handled case-insensitively.

Includes tests via the `WordPress.DB.RestrictedFunctions` sniff.