Move printing functions related functionality to dedicated `PrintingFunctionsTrait`
The printing function list is only used by a small set of sniffs, so are better placed in a dedicated trait.

The choice for a `trait` over a `class` is due to the `public` property allowing for adding additional functions to the list.

Moving both the base function list + the `public` property to the same trait will allow us to encapsulate all the functionality related to the use of this list in one place.

The `$printingFunctions` property, containing the base lists, has also been made `private`.

Checking whether or not something is a printing function should now be done by calling the `PrintingFunctionsTrait::is_printing_function()` method.

Related to 1465