Move formatting functions related functionality to dedicated `FormattingFunctionsHelper`
The formatting function list is only used by a small set of sniffs, so are better placed in a dedicated class.

The `$formattingFunctions` property has also been made `private static`.

Checking whether or not something is a formatting function should now be done by calling the `FormattingFunctionsHelper::is_formatting_function()` method.

Related to 1465