SanitizingFunctionsTrait: make sure function names are checked case-insensitively
These functions should be self-contained, so should not presume that the sniff has already lowercased the function name before passing it.

This fixes a bug as, in this case, the sniffs didn't actually lowercase the name before passing it to the Helper class methods, so the sniffs would throw false positives for non-lowercase function calls.

Tested by adjusting some pre-existing tests for the `ValidatedSanitizedInput` sniff.